move links to const file
