+++ packages/client/src/app.tsx
@@ -16,37 +16,7 @@ import {
   Register,
   registerLoader,
 } from './pages/index';
-
-export const links = {
-  root: {
-    path: '/',
-    title: 'Главная страница',
-  },
-  login: {
-    path: '/login',
-    title: 'Войти',
-  },
-  signup: {
-    path: '/signup',
-    title: 'Регистрация',
-  },
-  profile: {
-    path: '/profile',
-    title: 'Профиль',
-  },
-  leaderboard: {
-    path: '/leaderboard',
-    title: 'Лидеры',
-  },
-  forum: {
-    path: '/forum',
-    title: 'Форум',
-  },
-  game: {
-    path: '/game',
-    title: 'Игра',
-  },
-};
+import { links } from './utils/const';
 
 const router = createBrowserRouter([
   {
