+++ packages/client/src/utils/const.ts
@@ -1,0 +1,30 @@
+export const links = {
+  root: {
+    path: '/',
+    title: 'Главная',
+  },
+  login: {
+    path: '/login',
+    title: 'Войти',
+  },
+  signup: {
+    path: '/signup',
+    title: 'Регистрация',
+  },
+  profile: {
+    path: '/profile',
+    title: 'Профиль',
+  },
+  leaderboard: {
+    path: '/leaderboard',
+    title: 'Лидеры',
+  },
+  forum: {
+    path: '/forum',
+    title: 'Форум',
+  },
+  game: {
+    path: '/game',
+    title: 'Играть',
+  },
+};
